escape-changes insert intodb

+++ src/index.ts
@@ -1,4 +1,4 @@
-import { fabricToShema } from "./convert/fabricToShema";
+import { fabricToShema, queryJSON } from "./convert/fabricToShema";
 import { DesignTextRow } from "./interfaces/desginTextRow";
 import { ImageFactory } from "./factories/ImageFactory";
 import { TextFactory } from "./factories/TextFactory";
@@ -7,6 +7,7 @@ import { ElementTextRow } from "./interfaces/elementTextRow";
 import clipboardy from "clipboardy";
 import { jsonToSvg } from "./convert/jsonTo.svg";
 import fs from "fs";
+import { AbstractFactory } from "./factories/AbstractFactory";
 
 const config = {
   host: "127.0.0.1",
@@ -22,45 +23,17 @@ const promiseWrapper = async (desginID: number) => {
   const con = await mysql.createConnection(config);
 
   const ticketDesignQuery =
-    "SELECT * FROM ticketdesigns WHERE id=" + ticketDesginId;
+    "SELECT type, publicId FROM ticketdesigns WHERE id=" + ticketDesginId;
   const ticketElementsQuery =
     "SELECT * FROM ticketdesign_elements WHERE ticketdesign=" + ticketDesginId;
+
   const data = {
-    ticketDesign: (await con.query(ticketDesignQuery)) as DesignTextRow,
-    ticketElements: (await con.query(ticketElementsQuery)) as ElementTextRow,
+    ticketDesign: await con.query(ticketDesignQuery),
+    ticketElements: await con.query(ticketElementsQuery),
   };
+  con.end();
   return data;
 };
-
-class AbstractFactory {
-  static createNewDesign(
-    oldTicketElement: ElementTextRow,
-    ticketInfo: DesignTextRow
-  ) {
-    switch (oldTicketElement.type) {
-      case "text":
-        return TextFactory.convert(oldTicketElement);
-      case "text_v":
-        return TextFactory.convert(oldTicketElement);
-      case "required":
-        return TextFactory.convert(oldTicketElement);
-      case "barcode":
-        return ImageFactory.convert(oldTicketElement, ticketInfo);
-      case "barcode2":
-        return ImageFactory.convert(oldTicketElement, ticketInfo);
-      case "qrcode":
-        return ImageFactory.convert(oldTicketElement, ticketInfo);
-      case "picture":
-        return ImageFactory.convert(oldTicketElement, ticketInfo);
-      case "picture_eventlogo":
-        return ImageFactory.convert(oldTicketElement, ticketInfo);
-      case "picture_eventheader":
-        return ImageFactory.convert(oldTicketElement, ticketInfo);
-      case "picture_customerphoto":
-        return ImageFactory.convert(oldTicketElement, ticketInfo);
-    }
-  }
-}
 
 const fabricfy = async (oldElements: any) => {
   let fabricsElements = {
@@ -77,19 +50,26 @@ const fabricfy = async (oldElements: any) => {
       fabricsElements.objects.push(fabricElement);
     } catch (err) {}
   }
-  //clipboardy.writeSync(JSON.stringify(fabricsElements));
   return JSON.stringify(fabricsElements);
 };
 
 (async () => {
-  for (let i = 1; i < 2969; i++) {}
-  const oldElements = await promiseWrapper(2953);
-  const jsonFormat: JSON = JSON.parse(await fabricfy(oldElements));
-  const svg = jsonToSvg(jsonFormat);
-  console.log(svg);
-  // await fabricToShema(jsonFormat, svg, 2953);
-  if (fs.existsSync(`./src/assets/images/picture.jpeg`)) {
-    fs.unlinkSync(`./src/assets/images/picture.jpeg`);
+  for (let i = 1543; i <= 2968; i++) {
+    console.log(i);
+    const oldElements = await promiseWrapper(i);
+    if (
+      oldElements.ticketDesign[0].length === 0 ||
+      oldElements.ticketElements[0].length === 0
+    ) {
+      continue;
+    }
+
+    const jsonFormat = JSON.parse(await fabricfy(oldElements));
+    jsonToSvg(jsonFormat, i);
+    await queryJSON(jsonFormat, i);
+    if (fs.existsSync(`./src/assets/images/picture.jpeg`)) {
+      fs.unlinkSync(`./src/assets/images/picture.jpeg`);
+    }
   }
 })();
